fix syntax of version check for Pod::Readme
A string on a use line will be interpreted as an import argument, not a
version to check. In the case of Pod::Readme, this means that the
version will be entirely ignored. Future versions of perl are intending
to throw errors for arguments passed to an undefined import method.